pyproject.toml
[project]
name = "dungeon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce>=2.5.6",
]

[tool.ruff]
line-length = 120
